Create copyToClipboard helper
